Add SQLite3 DB, and partially set up MongoDB
Fixes: #8

I set up our SQLite3 db and added a Users table. We'll need to add bcrypt and a password column to this and whatnot, but the scaffolding is there. I also added some placeholder code for MongoDb to be added later down the line, but it's currently useless.

To run the migrations you run npx knex migrate:latest from the server dir.
To check on the db schema for our SQLite3 db I wrote a script located in scripts/check-schema.js. To run this you run ./scripts/check-schema.js from the server dir and it will dump all of our table names. I used this to ensure that the User table was up.

There is also code for a User model and userRoutes, but I have yet to test if these actually work. I need to add unit tests, and also hook it up to Drew's login page.

We're using Knex for our db config, btw. I also added some .env files for future use.