Start connecting up frontend to backend
I've began adding the routes for /register and /login for users. Currently this is hard to test from the actual frontend because we don't have a register option, only a login one, so I can't login without registering. But the general idea is there. I installed bcrypt, so our passwords should be getting hashed.