Embed stream iframe
So the OBS websocket concept is a no go. Apparently that feature only allows you to control the stream from the app, not actually watch it. So now we have to bounce the video like so:

Drone App on Phone -> Discord Video Call to Laptop -> Stream on OBS -> Youtube Stream -> Iframe Embed on App

We'll have to update the iframe every time we start a new stream, I suppose.