Revamp user login and logout,
Added auth provider, which handles the authentication state changing, logging in, and logging out. This state is shared with AuthProvider being a context provider to the HomePage and LoginForm.
Additionally, added a private route guard that checks for the current user. For example, if a user tries to visit /homepage , they will need to be logged in, if they arent they will be redirected to /login